Added Main Function Code and Edited my Find Function
I wrote both requirements for the first 2 assessments Task 1 where I needed to write print Hello world and The second assessment where I needed to test the rest of my functions and operators. While testing I had some output issues with my Find function so I edited it so that it is both more clear and it prints the right output.